изменила тип данных

+++ compl.py
@@ -4,13 +4,13 @@ def make_complex():
         if not (a.isdigit()):
             print('Ошибка. Повторите ввод')
             continue
-        a = float(a)
+        a = int(a)
         break
     while True:
         b = (input("Введите мнимую часть числа: "))
         if not (b.isdigit()):
             print('Ошибка. Повторите ввод')
             continue
-        b = float(b)
+        b = int(b)
         break
     return complex(a, b)
